moved fetchTrending to youtube_video_util
