+++ src/util/youtube_video_util.js
@@ -23,12 +23,24 @@ export const fetchDetails = (videoId, context) => {
 }
 
 export const fetchRelated = (videoId, context) => {
-    return fetch(`https://www.googleapis.com/youtube/v3/search?part=snippet&type=video&relatedToVideoId=${videoId}&key=${YT_API_KEY.publicDataKey}`)
-      .then(response => response.json())
-      .then(responseJson => {
-        context.setState({ vids: responseJson.items });
-      })
-      .catch(error => {
-        console.error(error);
-      })
-  }
+  return fetch(`https://www.googleapis.com/youtube/v3/search?part=snippet&type=video&relatedToVideoId=${videoId}&key=${YT_API_KEY.publicDataKey}`)
+    .then(response => response.json())
+    .then(responseJson => {
+      context.setState({ vids: responseJson.items });
+    })
+    .catch(error => {
+      console.error(error);
+    })
+}
+
+export const fetchTrending = () => {
+  let baseUrl = `https://www.googleapis.com/youtube/v3/videos`;
+  let part ='snippet';
+  let chart = 'mostPopular';
+  // let regionCode = '';
+  let maxResults = 25;
+
+  let fullUrl = `${baseUrl}?part=${part}&chart=${chart}&maxResults=${maxResults}&key=${YT_API_KEY.publicDataKey}`;
+
+  return fetch(fullUrl);
+};
